qr/requirements.txt
Pillow==5.4.1
qrcode==6.1
six==1.12.0
